📝 docs(README.md): update links and text The repository is now being used for the development of a series of articles on DukesCode about Solidity for beginners. The links to the articles have been updated to reflect this change. The motivation section has also been updated to provide more information about the series and its goals.
📝 chore(package.json): update description to mention DukesCode instead of Hashnode
The package.json file has been updated to reflect the new description of the repository. The description now mentions DukesCode instead of Hashnode as the platform where the Solidity for beginners series of articles is being published. This is a minor change that improves the accuracy of the repository's description.