📝 docs(README.md): update repository name and Hashnode link The repository name has been updated to reflect the new platform where the series of articles on Solidity for beginners is being published. The Hashnode link has been replaced with the new link to the DukesCode website.
📝 docs(README.md): update article link and description
The article link has been updated to reflect the correct URL where the article can be found. The description has also been updated to provide more clarity on the content of the article and the author's goals for the series.

📝 docs(README.md): update author's YouTube link
This commit updates the author's YouTube link in the README.md file to the correct URL. The previous link was broken, and the new link points to the author's YouTube channel.